Optimize the unsigned and uint64_t RNG code a bit
This helps a lot when gcc is being silly and can't figure out how to
inline for us and thereby do its memcmp(small-const) optimizations. It
also saves time by being willing to waste up to 3/7 bytes at the end
of a bufer.

On my laptop, this saves 18% for unsigned and 25% or uint64_t.