Remove the idx_step logic as too accident-prone
Having a counter which must only be advanced by a given step, and not
by one, is not a great design.  If we ever forgot to use the right
increment, we could generate the same part of the stream more than
once.

Instead, have the interface use a counter that increments by one per
call to ottery_prf.generate, and multiply it by the appropriate
multiple internally.