"Shallow" unit tests
These unit tests don't test the correctness of how libottery uses a
backend; they just check that some output is generated, that error
messages do/don't get generated as appropriate, and so on.

They also don't check (except in the weakest possible way) that the
output truly is random.

These tests are sufficient to cover every line in ottery.c and
ottery_osrng.c, except for the one line that calls abort().