Parameterize dump_bytes more
Since we want to be able to pass the output of libottery to a
statistical randomness test, we need to be able to say how to
configure it. It sure would be sad if we had random output for the
ottery_* functions but not the ottery_st_* functions, or something.