A better construction for forward secrecy.
DJB told me about this -- I should look up who it was that he said
came up with it before I attribute it to him.  (Nothing I'm about to
say constitutes endorsement by DJB of anything I'm doing here!)

The old one was "forward secure" with an M-byte window into the
past, where M was up to 16K or so.  This one should be properly
forward secure.

The new approach is to immediately re-key after we generate each
block of output, and to clear out each byte of the output buffer
that we yield to the user.  Now the forward secrecy is really
forward-secret.

I've bumped up the buffer size to get the overhead down to a maximum
of 5% or so.

The memset() of errant bytes doesn't seem to actually cost much in
performance.