Try to wipe the stack after calling the PRF.
Of course, doing this puts is in a state of sin, since C doesn't
really let you say "this value should be wiped, dammit!" in a
portable way.  We could abuse unions, but in my testing, this
approach seemed likeliest to work out okay.

Naturally, you shouldn't rely on this, but it might save somebody
some trouble sometime if their program has some foolish
stack-leaking bug.

It costs a few cycles per call to the crypto function; about 5% on
my desktop.