Select the SSSE3 flavor of chacha_krovetz.c correctly.
The additional CPU feature required by the fastest x86 version of
chacha_krovetz is SSSE3, not SSE3 (note the third S!) (specifically,
it wants _mm_shuffle_epi8, which is part of SSSE3 but not SSE3). The
build-time detection was wrong, causing both versions of the compiled
code to be identical, and the runtime detection was wrong, causing the
allegedly-SSSE3 code to be used on CPUs that didn't have that feature.
The build-time bug masked the runtime bug.

Many of the changes here are just s/\<sse3/ssse3/i.