Revamp the SIMD detection.
The convenience libraries containing specially-built variants of
chacha_krovetz.c now have generic names, as do all the macros and Makefile
conditionals.  The configure script knows how to detect the necessary
intrinsics for ARM and PPC as well as x86, and will arrange not to build
chacha_krovetz.c at all if the appropriate intrinsics are not available.