Use pthread_atfork to make sure that we never miss a fork.
This is pretty paranoid, to be honest: our getpid() checking should be
adequate except in the case where we fork, and then the parent dies,
and then the child forks again without invoking any ottery functions
in the meantime. In this case, we could wind up with the same PID as
the original process, which would keep us from noticing that the PID
had changed.

(This code still fails if we fork exactly UNSIGNED_MAX times *and*
wind up with the same PID as when we started.)